// src/components/ButtonComponent/types.ts
export enum BUTTON_VARIANTS {
  Pagination = 'pagination',
  PaginationSelected = 'pagination-selected',
  SliderButton = 'slider-button',
  SliderButtonDisable = 'slider-button-disable',
  ColorButton = 'color-button',
  ColorButtonSelected = 'color-button-selected',
  FavButton = 'fav-button',
  FavButtonSelected = 'fav-button-selected',
  Primary = 'primary',
  PrimarySelected = 'primary-selected',
  BackButton = 'back-button',
}

export enum BUTTON_SIZE {
  Small = 's',
  Medium = 'm',
  Large = 'l',
}